requirements.txt
mcp
azure-identity
openai
logging
ipykernel
mcp[cli]
pandas
pyarrow
mcp
yfinance
mcp[cli]
azure-kusto-data
azure-kusto-ingest
azure-identity
autogen
autogen-agentchat
autogen-ext
autogen-core
autogen-ext[mcp]
fastmcp